Microservices_3_Data_Visualisation/requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
protobuf==3.19.4
setuptools==44.0.0
six==1.16.0
flask
